split `knobs.dart` file
