+++ packages/widgetbook/lib/src/knobs/bool_knob.dart
@@ -1,6 +1,8 @@
 import 'package:provider/provider.dart';
-import 'package:widgetbook/src/fields/fields.dart';
-import 'package:widgetbook/src/knobs/knobs.dart';
+
+import '../fields/fields.dart';
+import 'knob.dart';
+import 'knobs_notifier.dart';
 
 class BoolKnob extends Knob<bool> {
   BoolKnob({
