+++ packages/widgetbook/lib/src/knobs/slider_knob.dart
@@ -1,7 +1,8 @@
 import 'package:provider/provider.dart';
-import 'package:widgetbook/src/knobs/knobs.dart';
 
 import '../fields/fields.dart';
+import 'knob.dart';
+import 'knobs_notifier.dart';
 
 class SliderKnob extends Knob<double> {
   SliderKnob({
