+++ packages/widgetbook/lib/src/knobs/text_knob.dart
@@ -1,7 +1,8 @@
 import 'package:provider/provider.dart';
-import 'package:widgetbook/src/knobs/knobs.dart';
 
 import '../fields/fields.dart';
+import 'knob.dart';
+import 'knobs_notifier.dart';
 
 class TextKnob extends Knob<String> {
   TextKnob({
