+++ packages/widgetbook/lib/widgetbook.dart
@@ -1,5 +1,5 @@
 export 'package:widgetbook/src/addons/addons.dart';
-export 'package:widgetbook/src/knobs/knobs.dart';
+export 'package:widgetbook/src/knobs/knobs_notifier.dart';
 export 'package:widgetbook/src/navigation/navigation.dart';
 export 'package:widgetbook/src/theming/widgetbook_theme.dart';
 export 'package:widgetbook/src/widgetbook.dart';
